requirements.txt
boto3
requests
feedparser
pyyaml
beautifulsoup4
